chore: update config
