+++ examples/refactor-react/src/index.tsx
@@ -5,6 +5,7 @@ import './index.css'
 
 
 const container = document.querySelector('#root');
-const root = createRoot(container);
+// biome-ignore lint/style/noNonNullAssertion: <explanation>
+const root = createRoot(container!);
 
 root.render(<Main />);
